Update: Loops, nested loops, and new workbooks
Added codes to allow DA Analysis workbook to show the stock year, total daily volume, and the return. Introduced code to allow for a nested for a header row, set initial volume to zero, established the number of rows to loop over, nested loop to analyze total volume of cells, starting price, and the ending price.